requirements.txt
certifi==2023.7.22
cffi==1.15.1
cryptography==41.0.4
pycparser==2.21
pylsqpack==0.3.17
PyNaCl==1.5.0
pyOpenSSL==23.2.0
